Add user management functionality
Users are those who are able to log into Network and manage the system.  This
functionality allows them to be created, edited and deleted.  These will generally
be representatives of the Network organization, i.e. Birmingham Education Foundation.

Users are not network participants or members.  Participants and members are those
the network is intended to help.  Users are those who are organizing the network.