Removed the signup functionality
The initial version of Network will not allow users to signup on their
own.  Administrators of the network will create all the user accounts.

The MEAN.JS application generator generated signup functionality that
needed to be disabled.  The functionality was not completely removed
but the ability to access it was.