Added signing in participants feature
When people participate in a network event, there attendance needs
to be recorded along with thier name and contact information.  This process
needs to flow smoothly and keep participants from be duplicated in the system
each time they attend an event.

New technical functionality introduced in this commit is the typeahead control.
This feature only implements a subset of the CRUD functionality and is accessible
from the NetworkEvent view page.  It does not have it's own place in the navigation.