Implement Member Model using Discriminator
Since a member is a participant with a few extra data fields, it makes sense to represent this inheritance.
I used the discriminator function of Mongoose to add the extra fields that Member requires on top of Participant.
I also changed the OtherNetworks and Groups to Arrays of Strings insteaf of nested documents.