Added match board functionality
Network actions are requests, offerings, declarations and Business Of The Network.
When participants make one of these actions, it is recorded on a match board with
all the participants that are interested in the action.  This commit allows those
actions and the people who match with them to be recorded.

https://github.com/edbirmingham/network/issues/26