Fix Recruiter Needed error on signin
What a participant is being signed into an event, if the participant
is also a member that was imported without the required fields then
an error will occur saving the participant that will prevent them
from being signed in.

To resolve the problem with a quick fix, the required member fields
have been made optional.  In the future we will need a more robust
solution.