devDependencies updates
I was setting up the network app on a new Mac, and ran into some issues when
running `npm install`, likely related to different versions of node and/or npm.

The node/npm version on my old machine were 0.12.7 and 2.11.3 respectively.

The node/npm versions on my new machine are 5.6.0 and 3.6.0 respectively.

I ran npm install, and verified that the app runs and the tests pass on both
the old and new machines with this change.