Fixed error converting participant to a member
When the member form was extracted from the create and edit pages, the
participant quick search was inadvertantly included in the form.  It
should have remained in the create page only.  The result was the
upgrading a participant that was searched for would fail to populate
the form fields and would cause a problem saving the member.