Added user admin authorization
In order to access administrative functionality, which is all the
current functionality, a user must be granted the admin permission.

In order to implement this permission simply for all the Network
functionality, the Devise controllers had to be included in the
Network code base so that the admin permission verification could
be skipped.