Added Letsencrypt for SSL/TLS certificate creation
In order to secure the Network application the Rails force_ssl
configuration option is not set.  This will redirect http requests
to https requests.

To support the force_ssl option, there must be an certificate.
Letsencrypt is being used to generate certificates on a monthly basis
in an automated way.  The benefits are primarily that certificates
are automatically renewed going forward and secondarily they are free.

See https://github.com/pixielabs/letsencrypt-rails-heroku and
https://letsencrypt.org/ for details.