Added mongo_id to the network event
Since some network events will be loaded from the previous mongodb
database, a column to put the mongodb database id in the postgresql
database will allow for correcting data that is mistakenly loaded.
Also, the row of the csv file for new events loaded from the spreadsheet
was added to the csv import script.