Fixed error on quick search
When a member search is done from the member index page the :q parameter
is the text typed in to the search box.  When a member search is done from
an autocomplete select box via xhr the :q parameter is a hash with the text
in a :terms key.  A recent change treated both scenarios as a string value
which resulted in an error in the autocomplete.