Fix the clip button for events
The clip button was previously not in a form.  When the copy event
functionality was added the event list table was put in a form.
This made the clip button submit the form since it was a submit type
of button.  The button is not a button type of button and it no longer
submits the form when clicked.