Add helper to simplify titles
I'm headed towards simplifying the layout... and I think that the header
system is the first piece that could be DRY'd up.

To that end, I'm borrowing a great method for cleaning up the page
titles from the blog in references, and implementing it in one place to
show y'all what it could do.

References:
- https://coderwall.com/p/a1pj7w/rails-page-titles-with-the-right-amount-of-magic