Added the ability to quickly add members to cohorts
Added a quick search for members with a add button to quickly add
members to a cohort.  Also added a table of members to the cohort
show page to be able to see the members currently in the cohort.

Added an "include" parameter to the member index json so that cohorts
can be added to a member search response to be able differntiate
duplicates.