Added attendance cohorts to events
Cohorts were being used to specify who should attend an event as well
as who was served at an event.  When a subset of a cohort being served
was invited to attend then specifying both in one field would not work.
Now there are separate cohort fields for specifying who should attend and
who is being served.  This allows for one teacher's students in an academy
to be invited and to also indicate that the event targets the academy.
Otherwise, it would appear that the whole academy was invited to the event.