Added a report for students by program
This is the first static report to be added to the application.  As a
result the navigation to reports needed to be added as well as the
report itself.  This report will need to be changed to use the star
schema once it is finished.  This report against the transactional
database was written to give Ed a timely report.