Fixed average attendance and unique attendees
The total row for average attendance was summing the average
attendance of each program which was very incorrect.  Now it averages
the average attendance of each program which is better but probably
not accurate.

The total row for unique attendees was summing the unique attendees
for each program which is not correct.  An individual attendee could
attend events in two different programs and get counted twice in the
total row.  This was fixed by doing a separate query to get the total
unique attendees overall.