Updated rails-html-sanitizer gem
https://nvd.nist.gov/vuln/detail/CVE-2018-3741

There is a possible XSS vulnerability in all rails-html-sanitizer gem
versions below 1.0.4 for Ruby. The gem allows non-whitelisted attributes
to be present in sanitized output when input with specially-crafted HTML
fragments, and these attributes can lead to an XSS attack on target
applications. This issue is similar to CVE-2018-8048 in Loofah.