Fixed JS problems related to messages
A message page load function call require a javascript library that was
only loaded on the messages new page.  As a result the standard rails
behavior of loading all coffee script files on every page caused a
javascript error.  In order to result this error the default behavior of
loading all files was disabled.  This caused other parts of the site to
quit working since the required coffee script was not being loaded.

To fix the problem, a condition was added to make sure the function was
defined before calling it.

Another option would be to move the required javascript library to the
application layout and load it everywhere.  This was not chosen to keep
from loading the library everywhere.