Minor UI improvements for messages
The message editor defaulted to center justification.  Now it defaults
to left justification.

The message show page was outputting the HTML on the page.  Added
html_safe to the show page so the body of the email would be formatted
like it was sent.  The HTML email body is sanitized before being
displayed so that malicious JS can't be executed by an email.