Improved wording on group member add
When adding members in a group, some fields like cohorts will be the
same for each member.  To make them easier to enter those fields will be
carried over from one member to the next.  Some wording at the top of
the page explained which fields would be copied over.  The grammar of
the sentence was improved as well as the order of the fields was changed
to match their order on the member form.